Refactor SearchConfiguration struct into its own file for improved organization. Remove the inline definition from DynamicListBuilder.swift and ensure consistent documentation across the new file. This change enhances code clarity and maintainability.
